[refactor/#63] activity 모달 구조 이동
